Move rule message templates to correct constants file
